+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Change Discord to be installed as a Flatpak.
 - Change Gwenview to be installed as a Flatpak.
 - Change KeePassXC to be installed as a Flatpak.
+- Change LibreOffice to be installed as a Flatpak.
 
 ## 3.0.1